first refactor attempt, need to track state with redux probably
